requirements.txt
pandas
requests
tweepy
textblob
scikit-learn
joblib
python-dotenv
